Change reduced_DOS to be a pair [eps, dens]
Now, all functions that take a DOS as input take [eps, dens], and
generate_DOS returns such a pair. This streamlines the use case where
you want to choose a good range for eps (based on temperature and E_f),
generate the DOS and plug that in to specific heat.